# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Implements the Mailio server based on https://mirs.mail.io/ specifications",
        "title": "Mailio Server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/domains": {
            "get": {
                "description": "Returns a list of all supported domains",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "List all domains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.UserDomain"
                            }
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "error creating server did",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/resolve/domain": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Identify if an email is a DIDcomm/Mailio or SMTP address, and resolve the DID if it’s from a Mailio server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messaging"
                ],
                "summary": "Resolve domain from email address (smtp or mailio)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "valid email address",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "force DNS update",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Domain"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "401": {
                        "description": "invalid signature or unauthorized to send messages",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "description": "Message is the error message",
                    "type": "string"
                }
            }
        },
...
        "types.Domain": {
            "type": "object",
            "properties": {
                "_deleted": {
                    "type": "boolean"
                },
                "_id": {
                    "type": "string"
                },
                "_rev": {
                    "description": "Rev is the revision number returned\n_Rev    string `json:\"_rev,omitempty\"`",
                    "type": "string"
                },
                "mailioDIDDomain": {
                    "description": "mailio domain (if supportsMailio)",
                    "type": "string"
                },
                "mailioPublicKey": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ok": {
                    "description": "_ID     string `json:\"_id,omitempty\"`",
                    "type": "boolean"
                },
                "supportsMailio": {
                    "description": "domain can be both (mailio and smtp)",
                    "type": "boolean"
                },
                "supportsStandardEmail": {
                    "description": "domain can be both (mailio and smtp)",
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "types.UserDomain": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}